+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/settings/AppInfoImprint.kt
@@ -3,7 +3,9 @@ package de.connect2x.trixnity.messenger.compose.view.settings
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalUriHandler
+import de.connect2x.trixnity.messenger.MatrixMessengerBaseConfiguration
 import de.connect2x.trixnity.messenger.compose.view.DI
 import de.connect2x.trixnity.messenger.compose.view.get
 import de.connect2x.trixnity.messenger.compose.view.i18n.I18nView
@@ -13,10 +15,12 @@ import de.connect2x.trixnity.messenger.compose.view.theme.components.AdaptiveDia
 import de.connect2x.trixnity.messenger.compose.view.theme.components.AdaptiveDialogScrollContent
 import de.connect2x.trixnity.messenger.compose.view.theme.components.ThemedAdaptiveDialog
 import de.connect2x.trixnity.messenger.compose.view.theme.messengerColors
-import de.connect2x.trixnity.messenger.MatrixMessengerBaseConfiguration
+import de.connect2x.trixnity.messenger.util.MatrixMarkdownFlavour
 import de.connect2x.trixnity.messenger.util.html.AutoLinkifyVisitor
 import de.connect2x.trixnity.messenger.util.html.HtmlVisitor
 import de.connect2x.trixnity.messenger.viewmodel.settings.AppInfoViewModel
+import org.intellij.markdown.html.HtmlGenerator
+import org.intellij.markdown.parser.MarkdownParser
 
 interface AppInfoImprintView {
     @Composable
@@ -40,7 +44,12 @@ internal fun Imprint(onClose: () -> Unit) {
     val i18n = DI.get<I18nView>()
     val imprint = DI.get<MatrixMessengerBaseConfiguration>().imprint ?: return
     val uriHandler = LocalUriHandler.current
-    val content = AutoLinkifyVisitor.process(HtmlVisitor.process(imprint))
+    val markdownFlavour = DI.current.get<MatrixMarkdownFlavour>()
+    val content = remember {
+        val parser = MarkdownParser(markdownFlavour)
+        val html = HtmlGenerator(imprint, parser.buildMarkdownTreeFromString(imprint), markdownFlavour).generateHtml()
+        AutoLinkifyVisitor.process(HtmlVisitor.process(html))
+    }
 
     ThemedAdaptiveDialog(onClose) {
         AdaptiveDialogHeader(onClose = onClose) {